requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.1
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blis==0.7.10
Brotli==1.1.0
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.1.4
confection==0.1.1
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.7
datasets==2.14.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distro==1.9.0
Django==4.2.7
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dspy-ai==2.4.5
effdet==0.4.1
exceptiongroup==1.1.2
executing==1.2.0
fastai==2.7.12
fastbook==0.0.29
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.18.0
fastprogress==1.0.3
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
graphviz==0.20.1
grpcio==1.62.2
grpcio-health-checking==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
iopath==0.1.10
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==7.8.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.5
jupyterlab_server==2.24.0
kiwisolver==1.4.4
langcodes==3.3.0
layoutparser==0.3.4
lxml==4.9.3
Mako==1.3.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
minio==7.2.7
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
mutagen==1.47.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.25.2
ollama==0.1.9
omegaconf==2.3.0
onnxruntime==1.15.1
openai==1.23.2
opencv-python==4.8.0.76
optuna==3.6.1
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandoc==2.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
pexpect==4.8.0
phonenumbers==8.13.34
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
plumbum==1.8.2
ply==3.11
portalocker==2.7.0
preshed==3.0.8
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyarrow-hotfix==0.6
pycocotools==2.0.6
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.15.1
PyJWT==2.8.0
pypandoc==1.11
pyparsing==3.0.9
pypdfium2==4.18.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.1
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.29
sqlparse==0.4.4
srsly==2.4.7
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
terminado==0.17.1
thinc==8.1.10
threadpoolctl==3.2.0
timm==0.9.5
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.9.0
unstructured==0.9.1
unstructured-inference==0.5.7
uri-template==1.3.0
urllib3==2.2.1
validators==0.28.1
wasabi==1.1.2
wcwidth==0.2.6
weaviate-client==4.6.3
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==12.0
widgetsnbextension==3.6.5
XlsxWriter==3.1.2
xmltodict==0.13.0
xxhash==3.3.0
yarl==1.9.2
yt-dlp==2023.12.30
zipp==3.16.2
